feat(view): add 'view' command (#27)
* feat(view): add a first draft of `tink view`

* fix(view): move to jsx, register command

Move to use JSX and use register command.

* fix(view): use common opts

* style(view): use jsx, improve components

* fix(view): fix error handling

* fix(view): fix dist-tag handling

* fix(view): use libnpm instead of npa

* fix(view): improve error handling

* feat(view): allow unlimited items for specified fields

Allow unlimited items when field is specified.

* feat(view): add packument info

Add packument info to outputted JSON, as is custom in `npm view`. It's a 
bit double if multiple versions are queried, but that's just how it is. 
Also allows packument-specific props to be queried.

* feat(view): allow dots in props

Allow dots (.) in props with a special prop syntax:

Data:
    { foo: {
      'bar.baz': 1,
      bar: { baz: 2 }
    } }

Query:
    foo[bar.baz] // -> 1

Useful to query versions, deps:

    time[0.10.1]
    dependencies[lodash.uniq]

* fix(view): properly show error stack

* feat(view): show deps as columns

Show dependencies in columns, matching the old behaviour.

* chore(git): clean merge conflict